requirements.txt
# Core dependencies
slack-bolt>=1.18.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.0

# AI Model providers
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0
transformers>=4.35.0
torch>=2.1.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# Logging and monitoring
structlog>=23.2.0
